Revert traversal shortcuts as well, since InkWell no longer has Focus nodes
