+++ packages/flutter/lib/src/widgets/focus_traversal.dart
@@ -2,12 +2,9 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-import 'dart:ui';
-
 import 'package:flutter/foundation.dart';
 import 'package:flutter/painting.dart';
 
-import 'actions.dart';
 import 'basic.dart';
 import 'binding.dart';
 import 'focus_manager.dart';
@@ -792,139 +789,4 @@ class DefaultFocusTraversal extends InheritedWidget {
 
   @override
   bool updateShouldNotify(DefaultFocusTraversal oldWidget) => policy != oldWidget.policy;
-}
-
-// A base class for all of the default actions that request focus for a node.
-class _RequestFocusActionBase extends Action {
-  _RequestFocusActionBase(LocalKey name) : super(name);
-
-  FocusNode _previousFocus;
-
-  @override
-  void invoke(FocusNode node, Intent tag) {
-    _previousFocus = WidgetsBinding.instance.focusManager.primaryFocus;
-    node.requestFocus();
-  }
-
-  @override
-  void debugFillProperties(DiagnosticPropertiesBuilder properties) {
-    super.debugFillProperties(properties);
-    properties.add(DiagnosticsProperty<FocusNode>('previous', _previousFocus));
-  }
-}
-
-/// An [Action] that requests the focus on the node it is invoked on.
-///
-/// This action can be used to request focus for a particular node, by calling
-/// [Action.invoke] like so:
-///
-/// ```dart
-/// Actions.invoke(context, const Intent(RequestFocusAction.key), focusNode: _focusNode);
-/// ```
-///
-/// Where the `_focusNode` is the node for which the focus will be requested.
-///
-/// The difference between requesting focus in this way versus calling
-/// [_focusNode.requestFocus] directly is that it will use the [Action]
-/// registered in the nearest [Actions] widget associated with [key] to make the
-/// request, rather than just requesting focus directly. This allows the action
-/// to have additional side effects, like logging, or undo and redo
-/// functionality.
-///
-/// However, this [RequestFocusAction] is the default action associated with the
-/// [key] in the [WidgetsApp], and it simply requests focus and has no side
-/// effects.
-class RequestFocusAction extends _RequestFocusActionBase {
-  /// Creates a [RequestFocusAction] with a fixed [key].
-  RequestFocusAction() : super(key);
-
-  /// The [LocalKey] that uniquely identifies this action to an [Intent].
-  static const LocalKey key = ValueKey<Type>(RequestFocusAction);
-
-  @override
-  void invoke(FocusNode node, Intent tag) {
-    super.invoke(node, tag);
-    node.requestFocus();
-  }
-}
-
-/// An [Action] that moves the focus to the next focusable node in the focus
-/// order.
-///
-/// This action is the default action registered for the [key], and by default
-/// is bound to the [LogicalKeyboardKey.tab] key in the [WidgetsApp].
-class NextFocusAction extends _RequestFocusActionBase {
-  /// Creates a [NextFocusAction] with a fixed [key];
-  NextFocusAction() : super(key);
-
-  /// The [LocalKey] that uniquely identifies this action to an [Intent].
-  static const LocalKey key = ValueKey<Type>(NextFocusAction);
-
-  @override
-  void invoke(FocusNode node, Intent tag) {
-    super.invoke(node, tag);
-    node.nextFocus();
-  }
-}
-
-/// An [Action] that moves the focus to the previous focusable node in the focus
-/// order.
-///
-/// This action is the default action registered for the [key], and by default
-/// is bound to a combination of the [LogicalKeyboardKey.tab] key and the
-/// [LogicalKeyboardKey.shift] key in the [WidgetsApp].
-class PreviousFocusAction extends _RequestFocusActionBase {
-  /// Creates a [PreviousFocusAction] with a fixed [key];
-  PreviousFocusAction() : super(key);
-
-  /// The [LocalKey] that uniquely identifies this action to an [Intent].
-  static const LocalKey key = ValueKey<Type>(PreviousFocusAction);
-
-  @override
-  void invoke(FocusNode node, Intent tag) {
-    super.invoke(node, tag);
-    node.previousFocus();
-  }
-}
-
-/// An [Intent] that represents moving to the next focusable node in the given
-/// [direction].
-///
-/// This is the [Intent] bound by default to the [LogicalKeyboardKey.arrowUp],
-/// [LogicalKeyboardKey.arrowDown], [LogicalKeyboardKey.arrowLeft], and
-/// [LogicalKeyboardKey.arrowRight] keys in the [WidgetsApp], with the
-/// appropriate associated directions.
-class DirectionalFocusIntent extends Intent {
-  /// Creates a [DirectionalFocusIntent] with a fixed [key], and the given
-  /// [direction].
-  const DirectionalFocusIntent(this.direction) : super(DirectionalFocusAction.key);
-
-  /// The direction in which to look for the next focusable node when the
-  /// associated [DirectionalFocusAction] is invoked.
-  final TraversalDirection direction;
-}
-
-/// An [Action] that moves the focus to the focusable node in the given
-/// [direction] configured by the associated [DirectionalFocusIntent].
-///
-/// This is the [Action] associated with the [key] and bound by default to the
-/// [LogicalKeyboardKey.arrowUp], [LogicalKeyboardKey.arrowDown],
-/// [LogicalKeyboardKey.arrowLeft], and [LogicalKeyboardKey.arrowRight] keys in
-/// the [WidgetsApp], with the appropriate associated directions.
-class DirectionalFocusAction extends _RequestFocusActionBase {
-  /// Creates a [DirectionalFocusAction] with a fixed [key];
-  DirectionalFocusAction() : super(key);
-
-  /// The [LocalKey] that uniquely identifies this action to [DirectionalFocusIntent].
-  static const LocalKey key = ValueKey<Type>(DirectionalFocusAction);
-
-  /// The direction in which to look for the next focusable node when invoked.
-  TraversalDirection direction;
-
-  @override
-  void invoke(FocusNode node, DirectionalFocusIntent tag) {
-    super.invoke(node, tag);
-    final DirectionalFocusIntent args = tag;
-    node.focusInDirection(args.direction);
-  }
 }
